Adicionar funções de CRUD para Estoque

+++ crm/views.py
@@ -4,8 +4,8 @@ from django.urls import reverse_lazy
 from django.http import HttpResponseForbidden
 from django.shortcuts import get_object_or_404, redirect
 
-from .forms import ProdutoForm
-from .models import Produto
+from .forms import EstoqueForm, ProdutoForm
+from .models import Produto, Estoque
 
 
 class ProdutoList(generic.ListView):
@@ -37,8 +37,43 @@ class ProdutoUpdate(VerificarSuperusuarioMixin, generic.UpdateView):
 
 
 def apagar_produto(request, pk):
-    if request.method == "POST":
+    usuario = request.user
+    autenticado = usuario.is_authenticated and usuario.is_superuser
+    if request.method == "POST" and autenticado:
         produto = get_object_or_404(Produto, id=pk)
         produto.delete()
         return redirect("produtos")
+    return HttpResponseForbidden()
+
+
+class EstoqueList(generic.ListView):
+    model = Estoque
+    context_object_name = "estoque"
+    template_name = "crm/listar_estoque.html"
+
+
+class EstoqueCreateForm(VerificarSuperusuarioMixin, generic.FormView):
+    form_class = EstoqueForm
+    template_name = "crm/criar_estoque.html"
+    success_url = reverse_lazy("listar_estoque")
+
+    def form_valid(self, form):
+        form.save()
+        return super().form_valid(form)
+
+
+class EstoqueUpdate(VerificarSuperusuarioMixin, generic.UpdateView):
+    model = Estoque
+    form_class = EstoqueForm
+    template_name = "crm/editar_estoque.html"
+    success_url = reverse_lazy("listar_estoque")
+
+
+def apagar_estoque(request, pk):
+    usuario = request.user
+    autenticado = usuario.is_authenticated and usuario.is_superuser
+    if request.method == "POST" and autenticado:
+        estoque = get_object_or_404(Estoque, id=pk)
+        estoque.delete()
+        return redirect("estoque")
     return HttpResponseForbidden()
